change damge to total attributes from initial attributes

+++ Hero/Services/DamageCalculator.cs
@@ -5,7 +5,7 @@ public class DamageCalculator
     public static int CalculateDamage(Weapon? equippedWeapon, int damagingAttribute)
     {
         double damageWeapon = equippedWeapon?.WeaponDamage ?? 1;
-        double damage = damageWeapon * (1 + (damagingAttribute / 100.0));
+        int damage = (int) (damageWeapon * (1 + (damagingAttribute / 100.0)));
         return (int)damage;
     }
 }
